newsreader: cleanup
